검색시 se에 주는 점수를 int로 바꿈. 이 값이 -1이면 score는 내부점수, 0 이상이면 해당값이 바로 점수가 된다.

+++ core/src/main/java/org/fastcatsearch/ir/search/AbstractPostingReader.java
@@ -6,10 +6,10 @@ public abstract class AbstractPostingReader implements PostingReader {
 
 	protected CharVector term;
 	protected int termPosition;
-	protected float weight;
+	protected int weight;
 	protected int documentCount;
 	
-	public AbstractPostingReader(CharVector term, int termPosition, float weight, int documentCount) {
+	public AbstractPostingReader(CharVector term, int termPosition, int weight, int documentCount) {
 		this.term = term;
 		this.termPosition = termPosition;
 		this.weight = weight;
@@ -17,7 +17,7 @@ public abstract class AbstractPostingReader implements PostingReader {
 	}
 
 	@Override
-	public float weight() {
+	public int weight() {
 		return weight;
 	}
